+++ docs/changelog.rst
@@ -5,3 +5,4 @@
 ##############
 
 * :feature:`-` Adds :py:func:`aiotk.mempipe`.
+* :feature:`-` Adds :py:class:`aiotk.UnixSocketServer`.